+++ src/main/java/fr/npagnucco/projetspring/repository/CategorieRepository.java
@@ -8,6 +8,6 @@ import fr.npagnucco.projetspring.model.Categorie;
  * Interface qui permet d'avoir accès aux méthodes de type save, findbyId, findall pour les catégories
  * @author nathan
  */
-public interface CategorieRepository extends JpaRepository<Categorie, Integer>{
+public interface CategorieRepository extends JpaRepository<Categorie, Long>{
 
 }
