+++ src/main/java/fr/npagnucco/projetspring/repository/PlatRepository.java
@@ -12,7 +12,7 @@ import fr.npagnucco.projetspring.model.Plat;
  * Interface qui permet d'avoir accès aux méthodes de type save, findbyId, findall pour les plats
  * @author nathan
  */
-public interface PlatRepository extends JpaRepository<Plat, Integer> {
+public interface PlatRepository extends JpaRepository<Plat, Long> {
     @Query("SELECT p FROM Plat p WHERE (:cid is NULL OR p.categorie.id = :cid) AND (:mincal IS NULL OR p.nb_calories>:mincal) AND (:maxcal IS NULL OR p.nb_calories<:maxcal)")
     Page<Plat> findByFiltre(@Param("cid") Integer cid,@Param("mincal") Integer mincal,@Param("maxcal") Integer maxcal, Pageable pageable);
